requirements.txt
setuptools==39.0.1
PyMySQL==0.9.3
pyhdfs==0.2.2
pyhive==0.6.1
pymssql==2.1.4
raven==6.10.0
yagmail==0.11.220
PyYAML==5.1.1
